fix: scope raw script content cache to the caller's authorization identity
The RAW_SCRIPT_CACHE in raw_script_by_path_internal is consulted before the
authed RLS transaction runs, and its key was built from the caller-supplied
cache_key query parameter with no identity component. Any workspace member
could replay a cache_key another caller had warmed and read folder-protected
script content they have no permission to see; entries also stayed valid after
a caller's permissions were revoked.

Prefix the content-cache key with auth_identity(&authed), the hash of the
caller's full authorization context, so a hit can only be returned to a caller
whose own authorized read populated it.

auth_identity moves from windmill-store's var_resource_cache to
windmill_common::db, next to the Authable trait it operates on, and is
re-exported from its old path.

Fixes WIN-2410
