+++ backend/windmill-common/src/db.rs
@@ -1,3 +1,4 @@
+use sha2::{Digest, Sha256};
 use sqlx::{Acquire, PgConnection, PgExecutor, Pool, Postgres, Transaction};
 
 use crate::audit::AuditAuthor;
@@ -117,6 +118,54 @@ impl Authable for Authed {
     fn username(&self) -> &str {
         &self.username
     }
+}
+
+/// Hash the caller's full authorization context into a stable identity string.
+///
+/// Caches that are consulted *before* the RLS query they short-circuit must key on this,
+/// so a hit can only ever be returned to a caller whose own authorized read populated it.
+///
+/// Email alone is **not** a sufficient scope: the same email can resolve to different
+/// effective permissions (`username`, groups, folders, scopes, admin/operator) through
+/// job- or owner-scoped tokens that share an email but carry a narrower `permissioned_as`.
+/// Every input that determines what the caller may read is folded in, mirroring
+/// `job_read_access_cache_key` in windmill-api, so a lower-privilege context can never
+/// reuse a higher-privilege context's cache entry. Variable-length fields are
+/// length-prefixed to keep the encoding injective.
+pub fn auth_identity<A: Authable + ?Sized>(authed: &A) -> String {
+    let mut hasher = Sha256::new();
+    let field = |hasher: &mut Sha256, bytes: &[u8]| {
+        hasher.update((bytes.len() as u32).to_be_bytes());
+        hasher.update(bytes);
+    };
+    hasher.update([authed.is_admin() as u8, authed.is_operator() as u8]);
+    field(&mut hasher, authed.email().as_bytes());
+    field(&mut hasher, authed.username().as_bytes());
+    let mut groups: Vec<&str> = authed.groups().iter().map(String::as_str).collect();
+    groups.sort_unstable();
+    hasher.update((groups.len() as u32).to_be_bytes());
+    for g in groups {
+        field(&mut hasher, g.as_bytes());
+    }
+    let mut folders: Vec<&str> = authed.folders().iter().map(|f| f.0.as_str()).collect();
+    folders.sort_unstable();
+    hasher.update((folders.len() as u32).to_be_bytes());
+    for f in folders {
+        field(&mut hasher, f.as_bytes());
+    }
+    match authed.scopes() {
+        // u32::MAX length-prefix marks "no scopes" so it can't collide with an empty list.
+        None => hasher.update(u32::MAX.to_be_bytes()),
+        Some(scopes) => {
+            let mut scopes: Vec<&str> = scopes.iter().map(String::as_str).collect();
+            scopes.sort_unstable();
+            hasher.update((scopes.len() as u32).to_be_bytes());
+            for s in scopes {
+                field(&mut hasher, s.as_bytes());
+            }
+        }
+    }
+    hex::encode(hasher.finalize())
 }
 
 lazy_static::lazy_static! {
@@ -288,5 +337,70 @@ impl<'b> DbExecutor<'b> for &'b mut PgConnection {
     }
     fn populate<'a>(&'a mut self) -> impl DbExecutor<'a> {
         &mut **self
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn base_authed() -> Authed {
+        Authed {
+            email: "alice@x.dev".to_string(),
+            username: "alice".to_string(),
+            is_admin: false,
+            is_operator: false,
+            groups: vec!["all".to_string()],
+            folders: vec![("shared".to_string(), false, false)],
+            scopes: None,
+            token_prefix: None,
+        }
+    }
+
+    // Email alone must NOT determine the cache identity: two contexts that share an email
+    // but resolve to different effective permissions must get distinct identities, so a
+    // lower-privilege context can never reuse a higher-privilege one's cached read.
+    #[test]
+    fn auth_identity_is_not_just_email() {
+        let base = auth_identity(&base_authed());
+
+        let mut more_folders = base_authed();
+        more_folders
+            .folders
+            .push(("secret".to_string(), false, false));
+        assert_ne!(base, auth_identity(&more_folders), "folders must matter");
+
+        let mut more_groups = base_authed();
+        more_groups.groups.push("devs".to_string());
+        assert_ne!(base, auth_identity(&more_groups), "groups must matter");
+
+        let mut other_user = base_authed();
+        other_user.username = "bob".to_string();
+        assert_ne!(base, auth_identity(&other_user), "username must matter");
+
+        let mut admin = base_authed();
+        admin.is_admin = true;
+        assert_ne!(base, auth_identity(&admin), "is_admin must matter");
+
+        let mut operator = base_authed();
+        operator.is_operator = true;
+        assert_ne!(base, auth_identity(&operator), "is_operator must matter");
+
+        let mut scoped = base_authed();
+        scoped.scopes = Some(vec!["resources:read:f/secret/x".to_string()]);
+        assert_ne!(base, auth_identity(&scoped), "scopes must matter");
+    }
+
+    // Identical authorization contexts must produce the same identity (so the same caller
+    // gets a cache hit), and ordering of groups/folders must not change the identity.
+    #[test]
+    fn auth_identity_is_stable_and_order_independent() {
+        assert_eq!(auth_identity(&base_authed()), auth_identity(&base_authed()));
+
+        let mut reordered = base_authed();
+        reordered.groups = vec!["all".to_string(), "devs".to_string()];
+        let mut other_order = base_authed();
+        other_order.groups = vec!["devs".to_string(), "all".to_string()];
+        assert_eq!(auth_identity(&reordered), auth_identity(&other_order));
     }
 }
